Split user listens response model classes to separate files
